requirements.txt
streamlit==1.30.0
mysqlclient==2.0.3
sqlalchemy==1.4.20
cryptography==3.4.7
pandas==1.3.0
numpy==1.21.0
pycryptodome==3.10.1